requirements.txt
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==0.8.11
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.1
spacy==3.7.2
plotly==5.17.0
